Add support for discover inner jars
It is quite common to embed jars in OSGi bundles and some are even ones
from maven central and then reexported to the outside. While this is
actually bad style it needs to be made visible and is part of a product
e.g a security scanner will possibly identify CVE by there.

This now adds a new option -process-bundle-classpath that enables
scanning for inner jars and if discovered add them as a subcomponent to
the SBOM.