Extract MavenDescriptor#createFromBytes
Currently there is one method to create a descriptor from properties,
but the code to discover it from the bytes (e.g. central search or
embedded maven infos) is inlined in the generic create method, this
makes it harder to reuse and understand.

This now extract the code into a method MavenDescriptor#createFromBytes
to make it easier to reuse.